initial pass at user mixins approach
Two tests are still failing.

I have a double presence check for Email is required `[ 'presence',
'presence', 'format.null' ]`

Also User ctor exports default Email model is returning Email instead
of email.
