+++ common/mixins/user/errors.js
@@ -1,0 +1,5 @@
+
+var debug = require('debug')('loopback:user:errors');
+
+module.exports.USERNAME_EMAIL_REQUIRED = 'USERNAME_EMAIL_REQUIRED';
+// module.exports.UserIdentity =
