+++ common/mixins/user/properties/username.js
@@ -1,0 +1,16 @@
+
+var debug = require('debug')('loopback:user:properties:username');
+
+module.exports = function(User, options) {
+  options = (options || {});
+  options.username = (options.username || {attribute: 'username', required: false});
+
+  debug('options.username', options.username);
+
+  User.defineProperty(options.username.attribute, {type: String, required: options.username.required});
+
+  User.once('setup', function(UserModel) {
+    UserModel.validatesUniquenessOf(options.username.attribute, {message: 'User already exists'});
+  });
+
+};
