+++ common/models/user.json
@@ -1,16 +1,6 @@
 {
   "name": "User",
   "properties": {
-    "realm": {
-      "type": "string"
-    },
-    "username": {
-      "type": "string"
-    },
-    "password": {
-      "type": "string",
-      "required": true
-    },
     "credentials": {
       "type": "object",
       "deprecated": true
@@ -19,15 +9,12 @@
       "type": "object",
       "deprecated": true
     },
-    "email": {
-      "type": "string",
-      "required": true
-    },
-    "emailVerified": "boolean",
-    "verificationToken": "string",
     "status": "string",
     "created": "date",
     "lastUpdated": "date"
+  },
+  "mixins": {
+    "EmailPasswordUser" : true
   },
   "hidden": ["password"],
   "acls": [
@@ -50,18 +37,6 @@
     },
     {
       "principalType": "ROLE",
-      "principalId": "$everyone",
-      "permission": "ALLOW",
-      "property": "login"
-    },
-    {
-      "principalType": "ROLE",
-      "principalId": "$everyone",
-      "permission": "ALLOW",
-      "property": "logout"
-    },
-    {
-      "principalType": "ROLE",
       "principalId": "$owner",
       "permission": "ALLOW",
       "property": "findById"
@@ -71,19 +46,6 @@
       "principalId": "$owner",
       "permission": "ALLOW",
       "property": "updateAttributes"
-    },
-    {
-      "principalType": "ROLE",
-      "principalId": "$everyone",
-      "permission": "ALLOW",
-      "property": "confirm"
-    },
-    {
-      "principalType": "ROLE",
-      "principalId": "$everyone",
-      "permission": "ALLOW",
-      "property": "resetPassword",
-      "accessType": "EXECUTE"
     }
   ],
   "relations": {
