+++ lib/builtin-models.js
@@ -1,6 +1,9 @@
 module.exports = function(registry) {
   // NOTE(bajtos) we must use static require() due to browserify limitations
 
+  registry.modelBuilder.mixins.define('EmailPasswordUser',
+    require('../common/mixins/email-password-user'));
+
   registry.Email = createModel(
     require('../common/models/email.json'),
     require('../common/models/email.js'));
